Move older server modules to Prototype/Legacy/
@noteed, hereon, I'll proceed to add Prototype/New and add the newer
server modules under that tree.
